backend/requirements.txt
blinker==1.9.0
click==8.3.0
colorama==0.4.6
Flask==3.1.2
flask-cors==6.0.1
flask-jwt-extended==4.6.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==8.3.0
python-dotenv==1.1.1
Werkzeug==3.1.3
PyJWT==2.8.0
argon2-cffi==23.1.0